+++ spec/json_web_token/jwt_spec.rb
@@ -76,21 +76,6 @@ module JsonWebToken
           end
         end
 
-        context 'w RS256 keys' do
-          let(:signing_key) { OpenSSL::PKey::RSA.generate(2048) }
-          let(:verifying_key) { signing_key.public_key }
-          let(:verify_options) { {alg: 'RS256', key: verifying_key} }
-          describe 'passing matching options' do
-            let(:sign_options) { {alg: 'RS256', key: signing_key} }
-            it_behaves_like 'does #verify'
-
-            it 'plausible' do
-              jwt = Jwt.sign(claims, sign_options)
-              expect(plausible_message_signature? jwt, 256).to be true
-            end
-          end
-        end
-
         context "w ES256 'alg' header parameter" do
           let(:algorithm) { 'ES256' }
           describe 'w passing a matching algorithm to #verify' do